refactor: move Format component out of Validations (#98)
