+++ src/components/shared/index.ts
@@ -1,6 +1,7 @@
 export * from './Caret';
 export * from './Description';
 export * from './Divider';
+export * from './Format';
 export * from './Property';
 export * from './Types';
 export * from './Validations';
